Update order of page transitions
